Suppress warnings from .bam readers, prevent integer overflow for bases counted which lead to negative values

+++ src/snpsvm/bamreading/HasBaseProgress.java
@@ -6,6 +6,6 @@ public interface HasBaseProgress {
 	 * Returns number of bases examined
 	 * @return
 	 */
-	public int getBasesCalled();
+	public long getBasesCalled();
 	
 }
